refactor: main button hooiks refactor
