+++ src/module/common/component/navigation/BottomBar/QuickActionsBottomBarItem/hooks/useQuickActionsBottomBarItem.tsx
@@ -1,40 +1,27 @@
 import { capitalize } from "@peersyst/react-utils";
 import { ResourceKey } from "i18next";
 import { ArrowSendIcon, ArrowReceiveIcon, SwapIcon, BuyIcon } from "icons";
-import useGetSwapLink from "module/common/hook/useGetSwapLink";
 import { useModalState } from "module/common/hook/useModalState";
-import useNavigation from "module/common/hook/useNavigation";
 import useTranslate from "module/common/hook/useTranslate";
 import { QRCodeIcon } from "module/common/icons/QRCodeIcon";
 import { QuickAction } from "module/home/component/feedback/QuickActionsModal.types";
 import ReceiveModal from "module/transaction/component/core/ReceiveModal/ReceiveModal";
 import SendModal from "module/transaction/component/core/SendModal/SendModal";
 import { Fragment } from "react";
-import { MainScreens } from "../../../MainNavigatorGroup/MainScreens";
-import { DAppScreens } from "module/dapp/navigator/DAppsNavigator.types";
-import { parseSignerDeepLinkData } from "module/signer/utils/parseSignerDeepLinkData";
-import useSignerModal from "module/signer/hooks/useSignerModal";
-import { useToast } from "@peersyst/react-native-components";
 import QrScanner from "module/common/component/input/QrScanner/QrScanner";
+import { useScanQuickAction } from "./useScanQuickAction";
+import { useSwapQuickAction } from "./useSwapQuickAction";
+import { useBuyQuickAction } from "./useBuyQuickAction";
 
 export function useQuickActionsBottomBarItem() {
     const { open, showModal, hideModal } = useModalState();
-    const translateError = useTranslate("error");
-    const { showSignerModal } = useSignerModal();
-    const { showToast } = useToast();
-    const { open: scanOpened, showModal: showScanModal, hideModal: hideScanModal } = useModalState();
+    const translate = useTranslate();
+
+    const { handleSwapPress } = useSwapQuickAction();
+    const { handleBuyPress } = useBuyQuickAction();
+    const { scanOpened, showScanModal, handleScan, hideScanModal } = useScanQuickAction();
     const { open: sendOpened, showModal: showSendModal, hideModal: hideSendModal } = useModalState();
     const { open: receiveOpened, showModal: showReceiveModal, hideModal: hideReceiveModal } = useModalState();
-    const navigate = useNavigation();
-
-    const uriSwap = useGetSwapLink();
-
-    function handleSwapPress(): void {
-        // (jordi): We need to cast the navigate.navigate to any because the React Navigation types are not working properly
-        navigate.navigate(MainScreens.DAPPS, { screen: DAppScreens.WEBVIEW, params: { url: uriSwap } } as any);
-    }
-
-    const translate = useTranslate();
 
     function getCapitalizedTranslation(key: ResourceKey): string {
         return capitalize(translate(key));
@@ -45,17 +32,6 @@ export function useQuickActionsBottomBarItem() {
             hideModal();
             callback();
         };
-    }
-
-    const handleScan = (data: string) => {
-        hideScanModal();
-        const signerData = parseSignerDeepLinkData(data);
-        if (!signerData) showToast(translateError("invalidSignerRequest"), { type: "error" });
-        else showSignerModal(signerData!.type, signerData!.id);
-    };
-
-    function handleBuyPress() {
-        navigate.navigate(MainScreens.FIAT_ORDERS);
     }
 
     const actions: QuickAction[] = [
